requirements.txt
Pillow==9.5.0
pytesseract==0.3.10
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyperclip==1.8.2
mss==9.0.1
numpy==1.24.3 